Move GenerateOSSMetricbeatModuleIncludeListGo in Metricbeat mage dir
