+++ dev-tools/mage/fields.go
@@ -146,26 +146,6 @@ func GenerateModuleIncludeListGo() error {
 	return GenerateIncludeListGo(nil, []string{"module"}, nil, "include/list.go", "", "include")
 }
 
-// GenerateOSSMetricbeatModuleIncludeListGo generates include/list_{suffix}.go files containing
-// a import statement for each module and dataset.
-func GenerateOSSMetricbeatModuleIncludeListGo() error {
-	err := GenerateIncludeListGo(
-		nil, []string{"module"},
-		[]string{"module/docker", "module/kubernetes"},
-		"include/list_common.go", "", "include")
-	if err != nil {
-		return err
-	}
-	err = GenerateIncludeListGo(
-		nil, []string{"module/docker", "module/kubernetes"},
-		nil,
-		"include/list_docker.go", "\n// +build linux darwin windows\n", "include")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 // GenerateIncludeListGo generates an include/list.go file containing imports
 // for the packages that match the paths (or globs) in importDirs (optional)
 // and moduleDirs (optional).
